refactor: 패키지를 행위자 기준에서 도메인 기준으로 재구성
