--- src/main/java/roomescape/admin/controller/AdminThemeController.java
@@ -1,42 +1,0 @@
-package roomescape.admin.controller;
-
-import java.net.URI;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import roomescape.admin.dto.AdminThemeRequest;
-import roomescape.admin.dto.AdminThemeResponse;
-import roomescape.admin.service.AdminThemeService;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("/admin/themes")
-public class AdminThemeController {
-    private final AdminThemeService adminThemeService;
-
-    public AdminThemeController(AdminThemeService adminThemeService) {
-        this.adminThemeService = adminThemeService;
-    }
-
-    @PostMapping
-    public ResponseEntity<AdminThemeResponse> createTheme(
-            @RequestBody AdminThemeRequest request
-            ) {
-        AdminThemeResponse response = adminThemeService.createTheme(request);
-        return ResponseEntity.created(URI.create("/admin/themes/" + response.id())).body(response);
-    }
-
-    @GetMapping
-    public ResponseEntity<List<AdminThemeResponse>> getThemes() {
-        List<AdminThemeResponse> responses = adminThemeService.getAllThemes();
-        return ResponseEntity.ok(responses);
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteTheme(
-            @PathVariable Long id
-    ) {
-        adminThemeService.deleteTheme(id);
-        return ResponseEntity.noContent().build();
-    }
-}
