--- src/main/java/roomescape/admin/dto/AdminThemeRequest.java
@@ -1,8 +1,0 @@
-package roomescape.admin.dto;
-
-public record AdminThemeRequest(
-        String name,
-        String description,
-        String imageUrl
-){
-}
