+++ src/main/java/roomescape/reservation/domain/Reservation.java
@@ -1,0 +1,36 @@
+package roomescape.reservation.domain;
+
+import java.time.LocalDate;
+import roomescape.reservationtime.domain.ReservationTime;
+import roomescape.theme.domain.Theme;
+
+public class Reservation {
+
+    private final Long id;
+    private final String name;
+    private final LocalDate date;
+    private final ReservationTime time;
+    private final Theme theme;
+
+    private Reservation(Long id, String name, LocalDate date, ReservationTime time, Theme theme) {
+        this.id = id;
+        this.name = name;
+        this.date = date;
+        this.time = time;
+        this.theme = theme;
+    }
+
+    public static Reservation of(Long id, String name, LocalDate date, ReservationTime time, Theme theme) {
+        return new Reservation(id, name, date, time, theme);
+    }
+
+    public static Reservation of(String name, LocalDate date, ReservationTime time, Theme theme) {
+        return new Reservation(null, name, date, time, theme);
+    }
+
+    public Long getId() { return id; }
+    public String getName() { return name; }
+    public LocalDate getDate() { return date; }
+    public ReservationTime getTime() { return time; }
+    public Theme getTheme() { return theme; }
+}
