+++ src/main/java/roomescape/reservation/dto/ReservationRequest.java
@@ -1,0 +1,11 @@
+package roomescape.reservation.dto;
+
+import java.time.LocalDate;
+
+public record ReservationRequest(
+        String name,
+        LocalDate date,
+        Long timeId,
+        Long themeId
+) {
+}
