+++ src/main/java/roomescape/reservation/dto/ReservationResponse.java
@@ -1,0 +1,22 @@
+package roomescape.reservation.dto;
+
+import java.time.LocalDate;
+import roomescape.reservation.domain.Reservation;
+import roomescape.reservationtime.dto.TimeResponse;
+
+public record ReservationResponse(
+        Long id,
+        String name,
+        LocalDate date,
+        TimeResponse time
+) {
+
+    public static ReservationResponse from(Reservation reservation) {
+        return new ReservationResponse(
+                reservation.getId(),
+                reservation.getName(),
+                reservation.getDate(),
+                TimeResponse.of(reservation.getTime())
+        );
+    }
+}
