+++ src/main/java/roomescape/reservationtime/controller/ReservationTimeController.java
@@ -1,0 +1,41 @@
+package roomescape.reservationtime.controller;
+
+import java.net.URI;
+import java.util.List;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.reservationtime.dto.TimeRequest;
+import roomescape.reservationtime.dto.TimeResponse;
+import roomescape.reservationtime.service.ReservationTimeService;
+
+@RestController
+public class ReservationTimeController {
+
+    private final ReservationTimeService reservationTimeService;
+
+    public ReservationTimeController(ReservationTimeService reservationTimeService) {
+        this.reservationTimeService = reservationTimeService;
+    }
+
+    @PostMapping("/times")
+    public ResponseEntity<TimeResponse> createTime(@RequestBody TimeRequest request) {
+        TimeResponse response = reservationTimeService.createTime(request);
+        return ResponseEntity.created(URI.create("/times/" + response.id())).body(response);
+    }
+
+    @GetMapping("/times")
+    public ResponseEntity<List<TimeResponse>> getTimes() {
+        return ResponseEntity.ok(reservationTimeService.getAllTimes());
+    }
+
+    @DeleteMapping("/times/{id}")
+    public ResponseEntity<Void> deleteTime(@PathVariable Long id) {
+        reservationTimeService.deleteById(id);
+        return ResponseEntity.noContent().build();
+    }
+}
