+++ src/main/java/roomescape/reservationtime/dto/TimeRequest.java
@@ -1,0 +1,9 @@
+package roomescape.reservationtime.dto;
+
+import java.time.LocalTime;
+
+public record TimeRequest(
+        LocalTime startAt,
+        LocalTime finishAt
+) {
+}
