+++ src/main/java/roomescape/reservationtime/dto/TimeResponse.java
@@ -1,0 +1,14 @@
+package roomescape.reservationtime.dto;
+
+import java.time.LocalTime;
+import roomescape.reservationtime.domain.ReservationTime;
+
+public record TimeResponse(
+        Long id,
+        LocalTime startAt
+) {
+
+    public static TimeResponse of(ReservationTime saved) {
+        return new TimeResponse(saved.getId(), saved.getStartAt());
+    }
+}
