+++ src/main/java/roomescape/theme/controller/ThemeController.java
@@ -1,0 +1,25 @@
+package roomescape.theme.controller;
+
+import java.util.List;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.theme.dto.ThemeResponse;
+import roomescape.theme.service.ThemeService;
+
+@RestController
+@RequestMapping("/themes")
+public class ThemeController {
+
+    private final ThemeService themeService;
+
+    public ThemeController(ThemeService themeService) {
+        this.themeService = themeService;
+    }
+
+    @GetMapping("/top")
+    public ResponseEntity<List<ThemeResponse>> getTopThemes() {
+        return ResponseEntity.ok(themeService.getTopThemes());
+    }
+}
