+++ src/main/java/roomescape/theme/dto/AdminThemeRequest.java
@@ -1,0 +1,8 @@
+package roomescape.theme.dto;
+
+public record AdminThemeRequest(
+        String name,
+        String description,
+        String imageUrl
+) {
+}
