+++ src/main/java/roomescape/theme/dto/ThemeResponse.java
@@ -1,0 +1,20 @@
+package roomescape.theme.dto;
+
+import roomescape.theme.domain.Theme;
+
+public record ThemeResponse(
+        Long id,
+        String name,
+        String description,
+        String imageUrl
+) {
+
+    public static ThemeResponse of(Theme theme) {
+        return new ThemeResponse(
+                theme.getId(),
+                theme.getName(),
+                theme.getDescription(),
+                theme.getImageUrl()
+        );
+    }
+}
