--- src/main/java/roomescape/user/controller/ReservationController.java
@@ -1,52 +1,0 @@
-package roomescape.user.controller;
-
-import java.time.LocalDate;
-import java.util.List;
-import java.net.URI;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-import roomescape.user.dto.ReservationRequest;
-import roomescape.user.dto.ReservationResponse;
-import roomescape.user.dto.TimeResponse;
-import roomescape.user.service.ReservationService;
-
-@RestController
-public class ReservationController {
-
-    private final ReservationService reservationService;
-
-    public ReservationController(ReservationService reservationService) {
-        this.reservationService = reservationService;
-    }
-
-    @PostMapping("/reservations")
-    public ResponseEntity<ReservationResponse> createReservation(
-        @RequestBody ReservationRequest request
-    ) {
-        ReservationResponse response = reservationService.createReservation(request);
-        return ResponseEntity.created(URI.create("/reservations/" + response.id())).body(response);
-    }
-
-    @GetMapping("/reservations")
-    public ResponseEntity<List<TimeResponse>> getReservations(
-        @RequestParam LocalDate date, @RequestParam Long themeId
-    ) {
-        List<TimeResponse> responses = reservationService.getReservations(date, themeId);
-        return ResponseEntity.ok(responses);
-    }
-
-    @DeleteMapping("/reservations/{id}")
-    public ResponseEntity<Void> deleteReservation(
-        @PathVariable Long id
-    ) {
-        reservationService.deleteReservation(id);
-        return ResponseEntity.noContent().build();
-    }
-
-}
