--- src/main/java/roomescape/user/dto/ReservationRequest.java
@@ -1,12 +1,0 @@
-package roomescape.user.dto;
-
-import java.time.LocalDate;
-
-public record ReservationRequest(
-    String name,
-    LocalDate date,
-    Long timeId,
-    Long themeId
-) {
-
-}
