--- src/main/java/roomescape/user/dto/ReservationTimeResponse.java
@@ -1,5 +1,0 @@
-package roomescape.user.dto;
-
-public record ReservationTimeResponse() {
-
-}
