--- src/main/java/roomescape/user/dto/ThemeResponse.java
@@ -1,20 +1,0 @@
-package roomescape.user.dto;
-
-import roomescape.admin.domain.Theme;
-
-public record ThemeResponse(
-    Long id,
-    String name,
-    String description,
-    String imageUrl
-) {
-
-    public static ThemeResponse of(Theme theme) {
-        return new ThemeResponse(
-            theme.getId(),
-            theme.getName(),
-            theme.getDescription(),
-            theme.getImageUrl()
-        );
-    }
-}
