--- src/main/java/roomescape/user/dto/TimeResponse.java
@@ -1,17 +1,0 @@
-package roomescape.user.dto;
-
-import java.time.LocalTime;
-import roomescape.user.domain.ReservationTime;
-
-public record TimeResponse(
-    Long id,
-    LocalTime startAt
-) {
-
-    public static TimeResponse of(ReservationTime saved) {
-        return new TimeResponse(
-            saved.getId(),
-            saved.getStartAt()
-        );
-    }
-}
